indexer/requirements.txt
mutagen==1.47.0
librosa==0.10.2
numpy==1.26.4
httpx==0.27.0
qdrant-client==1.9.1
sentence-transformers==3.0.1
torch==2.5.0
transformers==4.44.0
soundfile==0.12.1
fasttext-wheel==0.9.2
langdetect==1.0.9
beautifulsoup4==4.12.3
pyacoustid==1.2.2
musicbrainzngs==0.7.1
redis==5.0.7
rq==1.16.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.31
asyncpg==0.29.0
watchdog==4.0.1
pydantic-settings==2.3.4
python-dotenv==1.0.1
ruff==0.5.0
pytest==8.2.2
pytest-asyncio==0.23.7
rich==13.7.1
typer==0.12.3
click==8.1.8
